Version 2.0
Fixed issues latest version.

+++ Player.h
@@ -10,7 +10,7 @@ protected:
 		std::string Return_info(std::string);
 		bool is_modified;
 		int Return_numeric(std::string);
-		void Set_numeric(std::string type, float val);
+		void Set_numeric(std::string type, int val);
 		Player();
 		Player(int player_id, char* p_name, int highscore, bool mod);
 		Player(int player_id);
